refactor(api,transactional,ui,auth): collapse UserService, drop createResendClient, fix phantom deps
- apps/api: collapse UserService class to module-level async functions
  in apps/api/src/lib/users.ts; extract a single userSelect constant;
  drop the unused list() method; move the test file to lib/users.test.ts.
  Routes/v1/users.ts updated to call the new named functions.
- packages/transactional: drop createResendClient wrapper; call sites
  use new Resend(apiKey) directly. The existing apiKey guard in
  sendEmail and sendBatchEmails already covers the missing-key case.
- packages/ui: remove unused lucide-react dependency (apps/web and
  apps/landing already list it where it is actually imported).
- packages/auth: remove unused @tanstack/react-form devDependency.
