+++ packages/auth/package.json
@@ -22,7 +22,6 @@
   "devDependencies": {
     "@repo/config-vitest": "workspace:*",
     "@repo/typescript-config": "workspace:*",
-    "@tanstack/react-form": "^1.32.0",
     "@types/node": "^25.6.2",
     "@types/react": "^19.2.14",
     "react": "^19.2.6",
